+++ src/brasil/gov/tiles/tests/test_setup.py
@@ -31,6 +31,7 @@ TILES = [
     'standaloneheader',
     'video',
     'banner_rotativo',
+    'collective.cover.carousel',
 ]